// src/grammer/types.ts
export interface Node {
  type: string;
  loc: SourceLocation | null;
}

export interface SourceLocation {
  source: string | null;
  start: Position;
  end: Position;
}

export interface Position {
  line: number;
  column: number;
}

export interface Identifier extends Expression, Pattern {
  type: "Identifier";
  name: string;
}

export interface Literal extends Expression {
  type: "Literal";
  value: number | string | boolean;
  raw: string;
}

export interface RegExpLiteral extends Literal {
  regex: {
    pattern: string;
    flags: string;
  };
}

export interface Program extends Node {
  type: "Program";
  body: Array<Directive | Statement>;
  sourceType: string;
}

export interface Function extends Node {
  id: Identifier | null;
  params: Array<Pattern>;
  body: FunctionBody;
}

export interface Statement extends Node {}

export interface ExpressionStatement extends Statement {
  type: "ExpressionStatement";
  expression: Expression;
}

export interface Directive extends ExpressionStatement {
  expression: Literal;
  directive: string;
}

export interface BlockStatement extends Statement {
  type: "BlockStatement";
  body: Array<Statement>;
}

export interface FunctionBody extends BlockStatement {
  body: Array<Directive | Statement>;
}

export interface EmptyStatement extends Statement {
  type: "EmptyStatement";
}
// ...
export interface IfStatement extends Statement {
  type: "IfStatement";
  test: Expression;
  consequent: Statement;
  alternate: Statement | null;
}
// ...
export interface ForStatement extends Statement {
  type: "ForStatement";
  init: VariableDeclaration | Expression | null;
  test: Expression | null;
  update: Expression | null;
  body: Statement;
}

export interface ForInStatement extends Statement {
  type: "ForInStatement";
  left: VariableDeclaration | Pattern;
  right: Expression;
  body: Statement;
}

export interface Declaration extends Statement {}

// interface Statement
export interface VariableDeclaration extends Declaration {
  type: "VariableDeclaration";
  declarations: Array<VariableDeclarator>;
  kind: "var" | "let" | "const";
}

export interface VariableDeclarator extends Node {
  type: "VariableDeclarator";
  id: Pattern;
  init: Expression | null;
}

export interface Expression extends Node {}
// ...
export interface UpdateExpression extends Expression {
  type: "UpdateExpression";
  operator: UpdateOperator;
  argument: Expression;
  prefix: boolean;
}

export enum UpdateOperator {
  "++",
  "--",
}

export interface BinaryExpression extends Expression {
  type: "BinaryExpression";
  operator: BinaryOperator;
  left: Expression;
  right: Expression;
}

export enum BinaryOperator {
  "==",
  "!=",
  "===",
  "!==",
  "<",
  "<=",
  ">",
  ">=",
  "<<",
  ">>",
  ">>>",
  "+",
  "-",
  "*",
  "/",
  "%",
  "|",
  "^",
  "&",
  "in",
  "instanceof",
}
// ...
export interface MemberExpression extends Expression, Pattern {
  type: "MemberExpression";
  object: Expression;
  property: Expression;
  computed: boolean;
}

export interface ConditionalExpression extends Expression {
  type: "ConditionalExpression";
  test: Expression;
  alternate: Expression;
  consequent: Expression;
}

export interface CallExpression extends Expression {
  type: "CallExpression";
  callee: Expression;
  arguments: [Expression];
}

export interface NewExpression extends Expression {
  type: "NewExpression";
  callee: Expression;
  arguments: [Expression];
}

export interface SequenceExpression extends Expression {
  type: "SequenceExpression";
  expressions: [Expression];
}

export interface Pattern extends Node {}